Overloaded == was removed in micromachine 2.0.0
Commit b679d5a in soveran/micromachine removed the overloaded ==
operator for state machines.  I ran into this with some vagrant boxes
that required a reboot after upgrading virtual box tools.  Because ==
now compares the state machine object to a string, installation_ran? in
machine.rb always returns false.  So instead of running the reboot state,
vbguest fails with a loop detected since it proceeds to start instead:

    vagrant_vbguest.machine_loop_guard

See line:

https://github.com/dotless-de/vagrant-vbguest/blob/v0.12.0/lib/vagrant-vbguest/machine.rb#L60

To simulate a machine that needs reboot after installing the
tools, you can remove the ! in:

https://github.com/dotless-de/vagrant-vbguest/blob/v0.12.0/lib/vagrant-vbguest/machine.rb#L74