Added installer for Windows.
This commit was originally authored by fabriciocolombo and proposed as
pull request #129 on the vagrant-vbguest project. However, there have
been no updates from the author on this PR since 2014. This change
simply rebases fabriciocolombo's change to a more recent master, to
resolve merge conflicts.

This should resolve issue #65 in the vagrant-vbguest project.