(GH-298) Fix bug on EL7 where shared folders would not work
This patch adds the kernel-devel package to the list of dependencies.
This keeps previous behavior of installing kernel-devel-`uname -r` while
also installing kernel-devel. If a package is not found, yum will
continue to install the packages it does find, so there is no issue with
essentially listing it twice.