Fix check for running guest additions
At least with version 6.0.6 of the guest additions, the detection
mechanism does not work and always report them NOT to be running,
probably because they're still starting when the check occurs.

If the "vboxguest" driver is loaded and the "vboxsf" module exists
where the guest additions installer places it, it should be safe to
assume it is there and starting (except for the first time, which
triggers a VM reload anyway).